somw changes

+++ migrations/migrations/20251026201637_addStatusNotesToPatients.js
@@ -2,10 +2,17 @@
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.up = function(knex) {
-  return knex.schema.table("patients", function (table) {
-    table.string("status").defaultTo("Under Observation");
-    table.text("notes");
+exports.up = async function (knex) {
+  const hasStatus = await knex.schema.hasColumn("patients", "status");
+  const hasNotes = await knex.schema.hasColumn("patients", "notes");
+
+  return knex.schema.alterTable("patients", function (table) {
+    if (!hasStatus) {
+      table.string("status").defaultTo("Under Observation");
+    }
+    if (!hasNotes) {
+      table.text("notes");
+    }
   });
 };
 
@@ -13,9 +20,16 @@ exports.up = function(knex) {
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function(knex) {
-  return knex.schema.table("patients", function (table) {
-    table.dropColumn("status");
-    table.dropColumn("notes");
+exports.down = async function (knex) {
+  const hasStatus = await knex.schema.hasColumn("patients", "status");
+  const hasNotes = await knex.schema.hasColumn("patients", "notes");
+
+  return knex.schema.alterTable("patients", function (table) {
+    if (hasStatus) {
+      table.dropColumn("status");
+    }
+    if (hasNotes) {
+      table.dropColumn("notes");
+    }
   });
 };
